Move definition of DoubleWord into DoubleWord.swift
