+++ Sources/Atomics/Primitives/autogenerated/Primitives.native.swift
@@ -21,57 +21,9 @@
 #if ATOMICS_NATIVE_BUILTINS
 import Builtin
 
-#if (compiler(>=5.9) && _pointerBitWidth(_32)) || (compiler(<5.9) && (arch(i386) || arch(arm) || arch(arm64_32) || arch(wasm32)))
-@frozen
-@_alignment(8)
-public struct DoubleWord {
-  @usableFromInline
-  internal typealias _Builtin = Builtin.Int64
-
-  public var first: UInt
-  public var second: UInt
-
-  @inlinable @inline(__always)
-  public init(first: UInt, second: UInt) {
-    self.first = first
-    self.second = second
-  }
-}
-#else
-@frozen
-@_alignment(16)
-public struct DoubleWord {
-  @usableFromInline
-  internal typealias _Builtin = Builtin.Int128
-
-  public var first: UInt
-  public var second: UInt
-
-  @inlinable @inline(__always)
-  public init(first: UInt, second: UInt) {
-    self.first = first
-    self.second = second
-  }
-}
-#endif
-
-extension DoubleWord {
-  @_alwaysEmitIntoClient
-  @inline(__always)
-  internal init(_ builtin: _Builtin) {
-    self = unsafeBitCast(builtin, to: DoubleWord.self)
-  }
-
-  @_alwaysEmitIntoClient
-  @inline(__always)
-  internal var _value: _Builtin {
-    unsafeBitCast(self, to: _Builtin.self)
-  }
-}
-
 extension Bool {
   @_alwaysEmitIntoClient
-  @inline(__always)
+  @_transparent
   internal init(_ builtin: Builtin.Int1) {
     self = unsafeBitCast(builtin, to: Bool.self)
   }
